BasicMCP_With_Local_LLM_E2EProgram/requirements.txt
fastapi>=0.111,<0.116
uvicorn>=0.30,<0.31
aiohttp>=3.9,<3.11
PyYAML==6.0.2
ollama==0.3.3
mcp>=1.2.0
